[Fixes #160010580] create a tests module
